backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.129.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "sqlmodel>=0.0.33",
]
